feat: camera movement
- adds vertical, horizontal, and depthwise movement
- movement is tracked in terms of states which update on keys up/down.
  seems fairly robust but there may be bugs.
- movement is relative to the y rotation, while keeping x, z rots fixed
- some changes to data types to keep floats simpler (32 bit), if this
  causes problems we may revisit

+++ primitives/dt.odin
@@ -7,7 +7,7 @@ _EWMASmoothingFactor :: distinct f64
 EWMADt :: struct {
 	smoothing_factor: _EWMASmoothingFactor,
 	prev_tick:        Maybe(time.Tick),
-	dt:               Maybe(f64),
+	dt:               Maybe(f32),
 }
 
 EWMADtInitError :: enum {
@@ -47,7 +47,7 @@ _EWMADt_record_tick :: proc(dt: ^EWMADt, tick: time.Tick) {
 
 	prev_dt, has_prev_dt := dt.dt.?
 	if (!has_prev_dt) {
-		dt.dt = f64(time.tick_diff(prev_tick, tick)) * 0.000001 /* ms per ns */
+		dt.dt = f32(time.tick_diff(prev_tick, tick)) * 0.000001 /* ms per ns */
 		return
 	}
 
@@ -56,7 +56,7 @@ _EWMADt_record_tick :: proc(dt: ^EWMADt, tick: time.Tick) {
 }
 
 // Return duration in millis
-EWMADt_retrieve_millis :: proc(dt: ^EWMADt, default_fps: f64 = 60.0) -> f64 {
+EWMADt_retrieve_millis :: proc(dt: ^EWMADt, default_fps: f32 = 60.0) -> f32 {
 	duration, has_duration := dt.dt.?
 	if (!has_duration) {
 		// Return millis per fps
